requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
pdfplumber>=0.10.0
pillow>=10.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
google-generativeai>=0.6.0
pgvector>=0.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

